requirements.txt
requests
aiohttp
python-dotenv
tqdm
numpy
wandb
torch
torchvision
matplotlib